% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BerkeleyForestsAnalytics-package.R
\docType{package}
\name{BerkeleyForestsAnalytics-package}
\alias{BerkeleyForestsAnalytics}
\alias{BerkeleyForestsAnalytics-package}
\title{BerkeleyForestsAnalytics: Compute and Summarize Core Forest Metrics from Field Data}
\description{
A suite of open-source R functions designed to produce standard metrics for forest management and ecology from forest inventory data. The overarching goal is to minimize potential inconsistencies introduced by the algorithms used to compute and summarize core forest metrics. Learn more about the purpose of the package and the specific algorithms used in the package at \url{https://github.com/kearutherford/BerkeleyForestsAnalytics}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kearutherford/BerkeleyForestsAnalytics}
}

}
\author{
\strong{Maintainer}: Kea Rutherford \email{krutherford@berkeley.edu}

Authors:
\itemize{
  \item John Battles
  \item Danny Foster
}

Other contributors:
\itemize{
  \item The Regents of the University of California (Regents) [copyright holder]
}

}
\keyword{internal}
