% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_demo_data}
\alias{bio_demo_data}
\title{Data for biomass demonstrations}
\format{
A dataframe with 9 rows and 7 columns:
\describe{
\item{Forest}{Broader location or forest where the data were collected}
\item{Plot_id}{Plot in which the individual tree was measured}
\item{SPH}{Stems per hectare}
\item{Live}{Live (1) or dead (0)}
\item{Decay}{1-5 for standing dead trees. NA or 0 for live trees.}
\item{SPP}{Species of the individual tree, using four-letter species codes}
\item{DBH_CM}{Diameter at breast height in centimeters}
\item{HT_M}{Tree height in meters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
bio_demo_data
}
\description{
A fake dataset created for biomass demonstration purposes only
}
\keyword{datasets}
