% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsvb_formatting.R
\name{BiomassNSVB}
\alias{BiomassNSVB}
\title{BiomassNSVB}
\usage{
BiomassNSVB(
  data,
  sp_codes = "4letter",
  input_units = "metric",
  output_units = "metric",
  results = "by_plot"
)
}
\arguments{
\item{data}{describe}

\item{sp_codes}{describe}

\item{input_units}{describe}

\item{output_units}{describe}

\item{results}{describe}
}
\description{
Uses...
}
