% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_cwd_demo}
\alias{compilation_cwd_demo}
\title{Coarse woody debris data for compilation demonstrations}
\format{
A dataframe with 9 rows and 9 columns:
\describe{
\item{time}{year in which the data were collected}
\item{site}{broader location or forest where the data were collected}
\item{stratum}{stratum within site}
\item{plot}{plot within stratum}
\item{load_1000s_Mg_ha}{fuel load of sound 1000-hour fuels in megagrams per hectare}
\item{load_1000r_Mg_ha}{fuel load of rotten 1000-hour fuels in megagrams per hectare}
\item{load_cwd_Mg_ha}{total coarse woody debris fuel load in megagrams per hectare}
\item{sc_length_1000s}{slope-corrected transect length for sound 1000-hour fuels in meters}
\item{sc_length_1000r}{slope-corrected transect length for rotten 1000-hour fuels in meters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_cwd_demo
}
\description{
A fake dataset created for coarse woody debris compilation demonstration purposes only.
}
\keyword{datasets}
