% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_ffs_demo}
\alias{compilation_ffs_demo}
\title{Data for general Fire and Fire Surrogate demonstrations}
\format{
A dataframe with 9 rows and 8 columns:
\describe{
\item{time}{year in which the data were collected}
\item{trt_type}{treatment type (burn, thin, thin + burn, control)}
\item{site}{compartment}
\item{plot}{plot within compartment}
\item{sph}{stems per hectare}
\item{ba_m2_ha}{basal area in meters squared per hectare}
\item{qmd_cm}{quadratic mean diameter in centimeters}
\item{dbh_cm}{average diameter at breast height in centimeters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_ffs_demo
}
\description{
A fake dataset created for general compilation demonstration purposes only.
}
\keyword{datasets}
