% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_fpc_demo}
\alias{compilation_fpc_demo}
\title{FPC data for general simple random sampling demonstrations}
\format{
A dataframe with 2 rows and 3 columns:
\describe{
\item{site}{broader location or forest where the data were collected}
\item{N}{number of possible plots for the site}
\item{n}{number of plots measured}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_fpc_demo
}
\description{
A fake dataset created for general compilation demonstration purposes only.
}
\keyword{datasets}
