% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_fwd_demo}
\alias{compilation_fwd_demo}
\title{Fine woody debris data for compilation demonstrations}
\format{
A dataframe with 9 rows and 11 columns:
\describe{
\item{time}{year in which the data were collected}
\item{site}{broader location or forest where the data were collected}
\item{stratum}{stratum within site}
\item{plot}{plot within stratum}
\item{load_1h_Mg_ha}{fuel load of 1-hour fuels in megagrams per hectare}
\item{load_10h_Mg_ha}{fuel load of 10-hour fuels in megagrams per hectare}
\item{load_100h_Mg_ha}{fuel load of 100-hour fuels in megagrams per hectare}
\item{load_fwd_Mg_ha}{total fine woody debris fuel load in megagrams per hectare}
\item{sc_length_1h}{slope-corrected transect length for 1-hour fuels in meters}
\item{sc_length_10h}{slope-corrected transect length for 10-hour fuels in meters}
\item{sc_length_100h}{slope-corrected transect length for 100-hour fuels in meters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_fwd_demo
}
\description{
A fake dataset created for fine woody debris compilation demonstration purposes only.
}
\keyword{datasets}
