% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_srs_demo}
\alias{compilation_srs_demo}
\title{Data for general simple random sampling demonstrations}
\format{
A dataframe with 9 rows and 7 columns:
\describe{
\item{time}{year in which the data were collected}
\item{site}{broader location or forest where the data were collected}
\item{plot}{plot within site}
\item{sph}{stems per hectare}
\item{ba_m2_ha}{basal area in meters squared per hectare}
\item{qmd_cm}{quadratic mean diameter in centimeters}
\item{dbh_cm}{average diameter at breast height in centimeters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_srs_demo
}
\description{
A fake dataset created for general compilation demonstration purposes only.
}
\keyword{datasets}
