% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_srs_sp_demo}
\alias{compilation_srs_sp_demo}
\title{Data for general simple random sampling demonstrations}
\format{
A dataframe with 8 rows and 8 columns:
\describe{
\item{time}{year in which the data were collected}
\item{site}{broader location or forest where the data were collected}
\item{plot}{plot within site}
\item{species}{tree species}
\item{dominance}{relative basal area in percent}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_srs_sp_demo
}
\description{
A fake dataset created for general compilation demonstration purposes only.
}
\keyword{datasets}
