% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compilation_wt_demo}
\alias{compilation_wt_demo}
\title{Weight data for stratified random sampling demonstrations}
\format{
A dataframe with 4 rows and 3 columns:
\describe{
\item{site}{broader location or forest where the data were collected}
\item{stratum}{stratum within site}
\item{wh}{stratum weight}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
compilation_wt_demo
}
\description{
A fake dataset created for general compilation demonstration purposes only.
}
\keyword{datasets}
