% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwd_YS_demo}
\alias{cwd_YS_demo}
\title{Data for coarse woody debris demonstrations}
\format{
A dataframe with 12 rows and 8 columns:
\describe{
\item{time}{Year in which the data were collected}
\item{site}{Broader location or forest where the data were collected}
\item{plot}{Plot in which the individual transect was measured}
\item{transect}{Transect on which the 1000-hour fuel measurements were taken}
\item{length_1000h}{Length of the sampling transect for 1000-hour fuels in meters}
\item{slope}{Slope along the transect in percent}
\item{ssd_S}{Sum-of-squared-diameters for sound 1000-hour fuels}
\item{ssd_R}{Sum-of-squared-diameters for rotten 1000-hour fuels}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
cwd_YS_demo
}
\description{
A fake dataset created for coarse surface fuel demonstration purposes only
}
\keyword{datasets}
