% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{for_demo_data}
\alias{for_demo_data}
\title{Data for forest composition and structure demonstrations}
\format{
A dataframe with 9 rows and 7 columns:
\describe{
\item{site}{Broader location or forest where the data were collected}
\item{plot}{Plot in which the individual tree was measured}
\item{exp_factor}{Stems per hectare}
\item{status}{Live (1) or dead (0)}
\item{species}{Species of the individual tree, using four-letter species codes}
\item{dbh}{Diameter at breast height in centimeters}
\item{ht}{Tree height in meters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes
}
\usage{
for_demo_data
}
\description{
A fake dataset created for composition and structure demonstration purposes only
}
\keyword{datasets}
