% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_comp.R
\name{ForestComp}
\alias{ForestComp}
\title{ForestComp}
\usage{
ForestComp(data, relative = "BA", units = "metric")
}
\arguments{
\item{data}{A dataframe or tibble with the following columns: site, plot, exp_factor, status, species, and dbh. Each row must be an observation of an individual tree.}

\item{relative}{Not a variable (column) in the provided dataframe or tibble. Specifies whether forest composition should be measured as relative basal area or relative density. Must be set to either "BA" or "density". The default is set to "BA".}

\item{units}{Not a variable (column) in the provided dataframe or tibble. Specifies whether the dbh variable was measured using metric (centimeters) or imperial (inches) units. Must be set to either "metric" or "imperial". The default is set to "metric".}
}
\value{
A dataframe with the following columns:
\itemize{
\item site: as described above
\item plot: as described above
\item species: as described above
\item dominance: relative basal area (or relative density) in percent (\%)
}
}
\description{
Compiles forest composition at the plot level. Measured as relative basal area or relative density for live trees.
}
\examples{
ForestComp(data = for_demo_data,
           relative = "BA",
           units = "metric")

}
