% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_structure.R
\name{ForestStr}
\alias{ForestStr}
\title{ForestStr}
\usage{
ForestStr(data, units = "metric")
}
\arguments{
\item{data}{A dataframe or tibble with the following columns: site, plot, exp_factor, dbh, and ht (optional). Each row must be an observation of an individual tree.}

\item{units}{Not a variable (column) in the provided dataframe or tibble. Specifies (1) whether the dbh and ht variables were measured using metric (centimeters and meters) or imperial (inches and feet) units; (2) whether the expansion factor is in metric (stems per hectare) or imperial (stems per acre) units; and (3) whether results will be given in metric or imperial units. Must be set to either "metric" or "imperial". The default is set to "metric".}
}
\value{
A dataframe with the following columns:
\itemize{
\item site
\item plot
\item sph (or spa): stems per hectare (or stems per acre)
\item ba_m2_ha (or ba_ft2_ac): basal area in meters squared per hectare (or feet squared per acre).
\item qmd_cm (or qmd_in): quadratic mean diameter in centimeters (or inches). Weighted by the expansion factor.
\item dbh_cm (or dbh_in): average diameter at breast height in centimeters (or inches). Weighted by the expansion factor.
\item ht_m (or ht_ft): average height in meters (or feet) if ht argument was set. Weighted by the expansion factor.
}
}
\description{
Compiles forest structure at the plot level.
}
\examples{
ForestStr(data = for_demo_data,
          units = "metric")

}
