% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwd_demo}
\alias{fwd_demo}
\title{Data for fine woody debris demonstrations}
\format{
A dataframe with 12 rows and 11 columns:
\describe{
\item{time}{Year in which the data were collected}
\item{site}{Broader location or forest where the data were collected}
\item{plot}{Plot in which the individual transect was measured}
\item{transect}{Transect on which the fuel tallies were collected}
\item{count_1h}{Count of 1-hour fuels}
\item{count_10h}{Count of 10-hour fuels}
\item{count_100h}{Count of 100-hour fuels}
\item{length_1h}{Length of the sampling transect for 1-hour fuels in meters}
\item{length_10h}{Length of the sampling transect for 10-hour fuels in meters}
\item{length_100h}{Length of the sampling transect for 100-hour fuels in meters}
\item{slope}{Slope along the transect in percent}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
fwd_demo
}
\description{
A fake dataset created for fine surface fuel demonstration purposes only
}
\keyword{datasets}
