% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lit_duff_demo}
\alias{lit_duff_demo}
\title{Data for duff and litter demonstrations}
\format{
A dataframe with 24 rows and 6 columns:
\describe{
\item{time}{Year in which the data were collected}
\item{site}{Broader location or forest where the data were collected}
\item{plot}{Plot in which the individual transect was measured}
\item{transect}{Transect on which the duff/litter depths were measured}
\item{litter_depth}{Litter depth in centimeters}
\item{duff_depth}{Duff depth in centimeters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
lit_duff_demo
}
\description{
A fake dataset created for duff and litter fuel demonstration purposes only. Depths NOT already averaged together for each transect.
}
\keyword{datasets}
