% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overstory_demo}
\alias{overstory_demo}
\title{Overstory data for surface and ground fuel demonstrations}
\format{
A dataframe with 14 rows and 6 columns:
\describe{
\item{time}{Year in which the data were collected}
\item{site}{Broader location or forest where the data were collected}
\item{plot}{Plot in which the individual tree was measured}
\item{exp_factor}{Stems per hectare}
\item{species}{Species of the individual tree, using four-letter species codes}
\item{dbh}{Diameter at breast height in centimeters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes.
}
\usage{
overstory_demo
}
\description{
A fake dataset created for surface and ground fuel demonstration purposes only
}
\keyword{datasets}
