% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbiomass-package.R
\docType{package}
\name{Rbiomass-package}
\alias{Rbiomass}
\alias{Rbiomass-package}
\title{Rbiomass: Estimate Above-ground Stem, Bark, and Branch Tree Biomass}
\description{
Uses Forest Inventory and Analysis (FIA) Regional Biomass Equations to estimate above-ground stem, bark, and branch tree biomass. The package uses the California equation set and should not be used for data from other regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kearutherford/Rbiomass}
}

}
\author{
\strong{Maintainer}: Kea Rutherford \email{krutherford@berkeley.edu}

}
\keyword{internal}
