% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdi_demo_trees}
\alias{sdi_demo_trees}
\title{Tree data for stand density index demonstrations}
\format{
A dataframe with 10 rows and 5 columns:
\describe{
\item{site}{Forest where the data were collected}
\item{plot}{Plot in which the individual tree was measured}
\item{exp_factor}{Stems per hectare}
\item{status}{Live (1) or dead (0)}
\item{dbh}{Diameter at breast height in centimeters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes
}
\usage{
sdi_demo_trees
}
\description{
A fake dataset created for stand density index demonstration purposes only
}
\keyword{datasets}
