% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_density.R
\name{StandDensity}
\alias{StandDensity}
\title{StandDensity}
\usage{
StandDensity(
  tree_data,
  subsec_data,
  input_units = "metric",
  output_units = "metric"
)
}
\arguments{
\item{tree_data}{A dataframe or tibble with the following columns: site, plot, exp_factor, status, and dbh. Each row must be an observation of an individual tree.}

\item{subsec_data}{A dataframe or tibble with the following columns: site, plot (optional), and subsection.}

\item{input_units}{Not a variable (column) in the provided dataframe or tibble. Specifies (1) whether dbh was measured using metric (centimeters) or imperial (inches) units; and (2) whether the expansion factor is in metric (stems per hectare) or imperial (stems per acre) units. Must be set to either "metric" or "imperial". The default is set to "metric".}

\item{output_units}{Not a variable (column) in the provided dataframe or tibble. Specifies whether output SDI values will be given in metric (stems per hectare) or imperial (stems per acre) units. Must be set to either "metric" or "imperial". The default is set to "metric".}
}
\value{
A dataframe with the following columns:
\itemize{
\item site
\item plot
\item current_sdi_sph (or current_sdi_spa): current stand density index in stems per hectare (or stems per acre)
\item max_sdi_sph (or max_sdi_spa): maximum stand density index in stems per hectare (or stems per acre)
\item rel_density: relative density
}
}
\description{
Calculates stand density index.
}
\examples{
StandDensity(tree_data = sdi_demo_trees,
             subsec_data = sdi_demo_subs,
             input_units = "metric",
             output_units = "metric")

StandDensity(tree_data = sdi_demo_trees,
             subsec_data = sdi_demo_subs_plots,
             input_units = "metric",
             output_units = "metric")

}
