% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCBForestAnalytics-package.R
\docType{package}
\name{UCBForestAnalytics-package}
\alias{UCBForestAnalytics}
\alias{UCBForestAnalytics-package}
\title{UCBForestAnalytics: Produce Standard Metrics for Forest Management and Ecology from Field Inventory Data}
\description{
A suite of open-source R functions designed to produce standard metrics for forest management and ecology from field inventory data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kearutherford/UCBForestAnalytics}
}

}
\author{
\strong{Maintainer}: Kea Rutherford \email{krutherford@berkeley.edu}

Authors:
\itemize{
  \item John Battles
}

}
\keyword{internal}
