% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vign_fuels_5}
\alias{vign_fuels_5}
\title{Fuel data for vignette, version 5}
\format{
A dataframe with 236 rows and 16 columns:
\describe{
\item{time}{pre (pre-burn) or post (post-burn)}
\item{site}{compartment (60, 340, or 400)}
\item{plot}{plot in which the individual transect was measured}
\item{transect}{azimuth of transect on which the fuel data were collected}
\item{count_1h}{count of 1-hour fuels}
\item{count_10h}{count of 10-hour fuels}
\item{count_100h}{count of 100-hour fuels}
\item{length_1h}{length of the sampling transect for 1-hour fuels in meters}
\item{length_10h}{length of the sampling transect for 10-hour fuels in meters}
\item{length_100h}{length of the sampling transect for 100-hour fuels in meters}
\item{length_1000h}{length of the sampling transect for 1000-hour fuels in meters}
\item{ssd_S}{sum-of-squared-diameters for sound 1000-hour fuels}
\item{ssd_R}{sum-of-squared-diameters for rotten 1000-hour fuels}
\item{litter_depth}{litter depth in centimeters}
\item{duff_depth}{duff depth in centimeters}
\item{slope}{slope along the transect in percent}
}
}
\source{
Fire and Fire Surrogate Study, Stephens Lab, University of California, Berkeley
}
\usage{
vign_fuels_5
}
\description{
A clean dataset for demonstration purposes
}
\keyword{datasets}
