% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BerkeleyForestsAnalytics-package.R
\docType{package}
\name{BerkeleyForestsAnalytics-package}
\alias{BerkeleyForestsAnalytics}
\alias{BerkeleyForestsAnalytics-package}
\title{BerkeleyForestsAnalytics: Produce Standard Metrics for Forest Management and Ecology from Field Inventory Data}
\description{
A suite of open-source R functions designed to produce standard metrics for forest management and ecology from field inventory data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kearutherford/BerkeleyForestsAnalytics}
}

}
\author{
\strong{Maintainer}: Kea Rutherford \email{krutherford@berkeley.edu}

Authors:
\itemize{
  \item John Battles [copyright holder]
  \item Danny Foster
}

}
\keyword{internal}
