% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsvb_demo}
\alias{nsvb_demo}
\title{Data for NSVB framework biomass and carbon demonstrations}
\format{
A dataframe with 16 rows and 14 columns:
\describe{
\item{division}{Ecodivision in which the data were collected}
\item{province}{Province in which the data were collected}
\item{site}{Forest where the data were collected}
\item{plot}{Plot in which the individual tree was measured}
\item{exp_factor}{Stems per hectare}
\item{status}{Live (1) or dead (0)}
\item{decay_class}{1-5 for standing dead trees. NA for live trees.}
\item{Species}{Species of the individual tree, using four-letter species codes}
\item{dbh}{Diameter at breast height in centimeters}
\item{ht1}{Tree height 1 in meters}
\item{ht2}{Tree height 2 in meters}
\item{crown_ratio}{Live crown ratio}
\item{top}{Yes top (Y) or no top (N)}
\item{cull}{Percent wood cull}
}
}
\source{
Created by Kea Rutherford for demonstration purposes
}
\usage{
nsvb_demo
}
\description{
A fake dataset created for NSVB framework biomass and carbon demonstration purposes only
}
\keyword{datasets}
