% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdi_demo_subs}
\alias{sdi_demo_subs}
\title{Subsection data for stand density index demonstrations}
\format{
A dataframe with 2 rows and 2 columns:
\describe{
\item{site}{Forest where the data were collected}
\item{subsection}{Subsection in which the data were collected}
}
}
\source{
Created by Kea Rutherford for demonstration purposes
}
\usage{
sdi_demo_subs
}
\description{
A fake dataset created for stand density index demonstration purposes only
}
\keyword{datasets}
