% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdi_demo_subs_plots}
\alias{sdi_demo_subs_plots}
\title{Subsection data for stand density index demonstrations}
\format{
A dataframe with 4 rows and 3 columns:
\describe{
\item{site}{Forest where the data were collected}
\item{plot}{Plot in which the individual tree was measured}
\item{subsection}{Ecological Subsection in which the data were collected}
}
}
\source{
Created by Kea Rutherford for demonstration purposes
}
\usage{
sdi_demo_subs_plots
}
\description{
A fake dataset created for stand density index demonstration purposes only
}
\keyword{datasets}
