% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vign_trees_2}
\alias{vign_trees_2}
\title{Tree data for vignette, version 2}
\format{
A dataframe with 2250 rows and 10 columns:
\describe{
\item{id}{time:site combined}
\item{time}{pre (pre-burn) or post (post-burn)}
\item{site}{compartment (60, 340, or 400)}
\item{plot}{plot in which the individual tree was measured}
\item{exp_factor}{stems per hectare}
\item{status}{live (1) or dead (0)}
\item{decay}{1-5 for standing dead trees. 0 for live trees.}
\item{species}{Species of the individual tree, using four-letter species codes}
\item{dbh}{diameter at breast height in centimeters}
\item{ht}{tree height in meters}
}
}
\source{
Fire and Fire Surrogate Study, Stephens Lab, University of California, Berkeley
}
\usage{
vign_trees_2
}
\description{
A dataset with intentional errors for demonstration purposes
}
\keyword{datasets}
